INJIMOB-898 Remove Metro warnings (#1329)
* [INJIMOB-898] remove telemetry dependency from crypto utils


* [INJIMOB-898] remove vcMetadata dependency from utils file


* [INJIMOB-898] move common props key to constants


* [INJIMOB-898] remove vcUtils dependency from vcVerification


* [INJIMOB-898] remove profileIcon dependency from svg


* [INJIMOB-898] remove vcItemFeild dependecy from vcUtils


* [INJIMOB-898] remove vc dependecy from esignetVcItem and existing vcItem


* [INJIMOB-898] remove storage dependency from api.ts


* [INJIMOB-898] remove store dependency from storage


* [INJIMOB-898] remove commom props dependency from api.ts


* [INJIMOB-898] remove main.ts dependency from scanLayout


* [INJIMOB-898] move VcItemConatinerProfileImage to different component


* [INJIMOB-898] throw error when error is ENOENT and spell check


* [INJIMOB-898] rename types to routeTypes and edit imports


* [INJIMOB-898] move to idle state on tamperedVCs state


---------
