+++ machines/QrLoginMachine.ts
@@ -28,7 +28,7 @@ import {
 } from '../shared/telemetry/TelemetryUtils';
 import {TelemetryConstants} from '../shared/telemetry/TelemetryConstants';
 import {API_URLS} from '../shared/api';
-import getAllConfigurations from '../shared/commonprops/commonProps';
+import getAllConfigurations from '../shared/api';
 
 const model = createModel(
   {
