+++ machines/auth.ts
@@ -1,8 +1,6 @@
 import {assign, ContextFrom, EventFrom, send, StateFrom} from 'xstate';
 import {createModel} from 'xstate/lib/model';
-import getAllConfigurations, {
-  downloadModel,
-} from '../shared/commonprops/commonProps';
+import {downloadModel} from '../shared/api';
 import {AppServices} from '../shared/GlobalContext';
 import {StoreEvents, StoreResponseEvent} from './store';
 import {generateSecureRandom} from 'react-native-securerandom';
