+++ routes/main.ts
@@ -11,10 +11,6 @@ import {BOTTOM_TAB_ROUTES} from './routesConstants';
 import {HomeScreenLayout} from '../screens/HomeScreenLayout';
 import {Theme} from '../components/ui/styleUtils';
 
-export const changeTabBarVisible = (visible: string) => {
-  Theme.BottomTabBarStyle.tabBarStyle.display = visible;
-};
-
 const home: TabScreen = {
   name: BOTTOM_TAB_ROUTES.home,
   component: HomeScreenLayout,
@@ -49,24 +45,9 @@ mainRoutes.push(home);
 mainRoutes.push(share);
 mainRoutes.push(history);
 
-export type MainBottomTabParamList = {
-  home: undefined;
-  share: undefined;
-  history: undefined;
-};
-
 export interface TabScreen {
   name: string;
   icon: string;
   component: React.FC;
   options?: BottomTabNavigationOptions;
 }
-
-export type MainRouteProps = BottomTabScreenProps<
-  MainBottomTabParamList & RootStackParamList
->;
-
-export type HomeRouteProps = BottomTabScreenProps<
-  MainBottomTabParamList & RootStackParamList,
-  'home'
->;
