+++ screens/Request/RequestScreen.tsx
@@ -14,7 +14,7 @@ import {
   useFocusEffect,
   useNavigation,
 } from '@react-navigation/native';
-import {MainBottomTabParamList} from '../../routes/main';
+import {MainBottomTabParamList} from '../../routes/routeTypes';
 import {BOTTOM_TAB_ROUTES} from '../../routes/routesConstants';
 import {ProgressingModal} from '../../components/ProgressingModal';
 import {isIOS} from '../../shared/constants';
@@ -163,7 +163,7 @@ const SharingQR: React.FC<RequestScreenProps> = ({t, controller}) => {
 const StatusMessage: React.FC<RequestScreenProps> = ({t, controller}) => {
   return (
     controller.statusMessage !== '' && (
-      <Column testID='recievedCardStatus' elevation={1} padding="16 24">
+      <Column testID="recievedCardStatus" elevation={1} padding="16 24">
         <Text testID="receiveCardStatusMessage">
           {controller.statusMessage}
         </Text>
