+++ screens/Scan/ScanLayoutController.ts
@@ -6,7 +6,7 @@ import {
   MessageOverlayProps,
   VCSharingErrorStatusProps,
 } from '../../components/MessageOverlay';
-import {MainBottomTabParamList, changeTabBarVisible} from '../../routes/main';
+import {MainBottomTabParamList} from '../../routes/routeTypes';
 import {GlobalContext} from '../../shared/GlobalContext';
 import {
   selectIsConnecting,
@@ -36,10 +36,15 @@ import {
 import {ScanEvents} from '../../machines/bleShare/scan/scanMachine';
 import {BOTTOM_TAB_ROUTES, SCAN_ROUTES} from '../../routes/routesConstants';
 import {ScanStackParamList} from '../../routes/routesConstants';
+import {Theme} from '../../components/ui/styleUtils';
 
 type ScanLayoutNavigation = NavigationProp<
   ScanStackParamList & MainBottomTabParamList
 >;
+
+const changeTabBarVisible = (visible: string) => {
+  Theme.BottomTabBarStyle.tabBarStyle.display = visible;
+};
 
 // TODO: refactor
 // eslint-disable-next-line sonarjs/cognitive-complexity
