+++ shared/api.ts
@@ -1,9 +1,21 @@
 import {request} from './request';
-import {API_CACHED_STORAGE_KEYS} from './storage';
-import {COMMON_PROPS_KEY} from './commonprops/commonProps';
+import {
+  API_CACHED_STORAGE_KEYS,
+  COMMON_PROPS_KEY,
+  changeCrendetialRegistry,
+} from './constants';
 import {INITIAL_CONFIG} from './InitialConfig';
 import Keychain from 'react-native-keychain';
 import {getItem, setItem} from '../machines/store';
+import {faceMatchConfig} from './commonUtil';
+import {configure} from '@iriscan/biometric-sdk-react-native';
+import {
+  getErrorEventData,
+  getImpressionEventData,
+  sendErrorEvent,
+  sendImpressionEvent,
+} from './telemetry/TelemetryUtils';
+import {TelemetryConstants} from './telemetry/TelemetryConstants';
 
 export const API_URLS: ApiUrls = {
   issuersList: {
@@ -261,6 +273,60 @@ async function generateCacheAPIFunctionWithAPIPreference(
   }
 }
 
+export default async function getAllConfigurations(
+  host = undefined,
+  isCachePreferred = true,
+) {
+  host && changeCrendetialRegistry(host);
+  return await CACHED_API.getAllProperties(isCachePreferred);
+}
+
+export async function downloadModel() {
+  try {
+    console.log('restart Face model init');
+    var injiProp = await getAllConfigurations();
+    const maxRetryStr = injiProp.modelDownloadMaxRetry;
+    const maxRetry = parseInt(maxRetryStr);
+    const resp: string = injiProp != null ? injiProp.faceSdkModelUrl : null;
+
+    if (resp != null) {
+      for (let counter = 0; counter < maxRetry; counter++) {
+        let config = faceMatchConfig(resp);
+        var result = await configure(config);
+        console.log('model download result is = ' + result);
+        if (result) {
+          sendImpressionEvent(
+            getImpressionEventData(
+              TelemetryConstants.FlowType.faceModelInit,
+              TelemetryConstants.Screens.home,
+              {status: TelemetryConstants.EndEventStatus.success},
+            ),
+          );
+          break;
+        } else if (!result && counter === maxRetry - 1) {
+          sendErrorEvent(
+            getErrorEventData(
+              TelemetryConstants.FlowType.faceModelInit,
+              TelemetryConstants.ErrorId.failure,
+              TelemetryConstants.ErrorMessage.faceModelInitFailed,
+            ),
+          );
+        }
+      }
+    }
+  } catch (error) {
+    sendErrorEvent(
+      getErrorEventData(
+        TelemetryConstants.FlowType.faceModelInit,
+        TelemetryConstants.ErrorId.failure,
+        TelemetryConstants.ErrorMessage.faceModelInitFailed,
+        error,
+      ),
+    );
+    console.log(error);
+  }
+}
+
 type Api_Params = {
   method: 'GET' | 'POST' | 'PATCH'; // Define the HTTP methods
   buildURL: (param?: string) => `/${string}`; // Define the buildURL function signature
@@ -287,3 +353,8 @@ type ApiUrls = {
   sendConsent: Api_Params;
   googleAccountProfileInfo: Api_Params;
 };
+
+export interface DownloadProps {
+  maxDownloadLimit: number;
+  downloadInterval: number;
+}
