+++ shared/telemetry/TelemetryConstants.js
@@ -16,6 +16,8 @@ export const TelemetryConstants = {
     dataRestore: 'Data Restore',
     decryption: 'Decryption',
     dataBackupAndRestoreSetup: 'Data Backup & Restore Setup',
+    remove: 'remove VC',
+    removeVcMetadata: 'VC metadata removed',
   }),
 
   EndEventStatus: Object.freeze({
